Cleanup and add to main
