+++ src/index.tsx
@@ -12,9 +12,8 @@ import { ThemeProvider } from '@mui/material/styles';
 import { theme } from './common/Theme';
 import NotFound from './routes/NotFound';
 import { Toolbar } from '@mui/material';
+import Main from './routes/Main';
 import Projects from './routes/Projects';
-import Temp from './Temp';
-import MainAnimation from './routes/MainAnimation';
 
 ReactDOM.render(
   <React.StrictMode>
@@ -32,8 +31,7 @@ ReactDOM.render(
           <Toolbar /> {/* For spacing */}
           <main>
             <Routes>
-              <Route path="/" element={<Temp />} />
-              <Route path="animation" element={<MainAnimation />} />
+              <Route path="/" element={<Main />} />
               <Route path="dev" element={<Projects />} />
               <Route path="notes" element={<NotesIndex />} />
               <Route path="notes/:courseId" element={<NotesCourseMainRoute />} />
