+++ src/polygon/PolygonAnimation.scss
@@ -1,4 +1,5 @@
 svg.svg-grid {
+  content-visibility: auto;
   circle {
     fill: var(--svg-color);
     r: 0.1;
